add digits option
